fix(symbol): deterministic branch reporting, u32 polygon-vertex handles, expanded-submenu clamp
Below-cap findings from the same review pass:

- chain.rs's validate_topology iterated its endpoint-cluster HashMap
  in (unstable) hash order, so which node's point got reported when
  several independently qualified as Branching — and which of two
  dangling nodes landed in OpenChain's ends[0] vs ends[1] — wasn't
  guaranteed stable across runs. Both loops now walk a sorted Vec of
  root ids instead.
- GraphicHandle::PolygonVertex and its message-layer mirror
  GraphicHandleMsg::PolygonVertex are now u32, not u16: a u16 index
  truncates/wraps past 65535 vertices, colliding two different
  vertices onto the same handle id. The one cast site is
  debug_assert-guarded against ever needing to saturate.
- The symbol context menu's screen-edge clamp now accounts for the
  Place ▸ submenu's ~7 extra in-place rows when it's expanded, so an
  expanded card no longer under-estimates its own footprint and run
  off the bottom edge near the window boundary.

+++ crates/signex-app/src/library/editor/symbol/state/mod.rs
@@ -123,7 +123,11 @@ pub enum GraphicHandle {
     /// Polygon vertex at the given index into `vertices`. A selected
     /// polygon shows one of these per vertex; dragging one moves
     /// only that vertex (mirrors `LineEndpoint`'s per-point drag).
-    PolygonVertex(u16),
+    /// `u32`, not `u16`: a `u16` truncates/wraps for a polygon past
+    /// 65535 vertices, colliding two different vertices onto the same
+    /// handle id (see `graphic_handles`'s cast, which is `debug_assert`-
+    /// guarded against ever needing to saturate).
+    PolygonVertex(u32),
 }
 
 /// Map a [`GraphicHandle`] to the mouse cursor that should be shown
